requirements.txt
langchain
langchain-community
chromadb
streamlit
pysrt==1.1.2
python-dotenv
langchain-google-genai>=0.1.3
google-generativeai>=0.8.5
fastapi
uvicorn
requests
pydantic